Version final
Version final

+++ routes/dashboard.php
@@ -1,8 +1,11 @@
 
 <?php
 
+use App\Http\Controllers\CallsheetsController;
+use App\Http\Controllers\ProjectAssignmentController;
 use App\Http\Controllers\ProjectController;
 use App\Http\Controllers\ProjectInvitationController;
+use App\Http\Controllers\StatsController;
 use App\Http\Middleware\VerifyProjectOwnership;
 use Illuminate\Support\Facades\Route;
 use Inertia\Inertia;
@@ -20,14 +23,42 @@ Route::middleware(['auth', 'verified'])->group(function () {
         Route::get('/other-projects', [ProjectController::class, 'otherProjects'])->name('otherprojects');
         Route::get('/create', [ProjectController::class, 'create'])->name('create');
         Route::post('/', [ProjectController::class, 'store'])->name('store');
+        Route::get('/guest/{id}', [ProjectController::class, 'showGuest'])->name('guest.show'); //metodo pendiente
+
 
         // Rutas con validación de propietario
-        Route::middleware(VerifyProjectOwnership::class)->group(function () {
-            Route::get('/{id}', [ProjectController::class, 'show'])->name('show');
-            Route::get('/{id}/edit', [ProjectController::class, 'edit'])->name('edit');
-            Route::put('/{id}', [ProjectController::class, 'update'])->name('update');
-            Route::delete('/{id}', [ProjectController::class, 'destroy'])->name('destroy');
+        Route::middleware(VerifyProjectOwnership::class)->prefix('admin')->group(function () {
+            Route::get('/{project_id}', [ProjectController::class, 'showAdmin'])->name('admin.show');
+            Route::get('/{project_id}/edit', [ProjectController::class, 'edit'])->name('admin.edit');
+            Route::put('/{id}', [ProjectController::class, 'update'])->name('admin.update');
+            Route::delete('/{id}', [ProjectController::class, 'destroy'])->name('admin.destroy');
+
+            //participants
+            Route::get('/participant/{assignment_id}', [ProjectAssignmentController::class, 'editParticipant'])->name('admin.editParticipant');
+
+            Route::post('/participant/assign_role', [ProjectAssignmentController::class, 'assignRole'])->name('admin.assignRole');
+
+            //callsheets
+
+            Route::get('/{project_id}/callsheets', [CallsheetsController::class, 'index'])->name('admin.callsheets');
+            Route::post('/{project_id}/callsheets/store', [CallsheetsController::class, 'store'])->name('admin.callsheets.store');
+
+            Route::get('{project_id}/callsheets/{callsheet_id}', [CallsheetsController::class, 'show'])->name('admin.callsheets.show');
+
+            //callsheetCast
+            Route::post('callsheets/addCast', [CallsheetsController::class, 'addCast'])->name('admin.callsheets.cast.add');
+
+
+            //events
+            Route::post('callsheets/addEvent', [CallsheetsController::class, 'addEvent'])->name('admin.callsheets.event.add');
+
+            //core
+
+            Route::get('/stats/{project_id}', [StatsController::class, 'stats'])->name('admin.stats');
+
+            Route::post('/stats/resolveEvent', [StatsController::class, 'resolveEvent'])->name('admin.stats.event.resolve');
         });
+
 
         // Rutas adicionales relacionadas con proyectos
         Route::post('/invite', [ProjectInvitationController::class, 'invite'])->name('invite');
